+++ crates/compiler/src/lowering/components.rs
@@ -3,15 +3,20 @@
 use std::{
     borrow::Borrow,
     collections::HashMap,
+    error::Error,
     fmt::{self, Display, Formatter},
     hash::Hash,
     ops::Deref,
     path::PathBuf,
+    str::FromStr,
     sync::Arc,
 };
 use hotg_rune_core::Shape;
+use indexmap::IndexMap;
 use legion::Entity;
-use crate::parse::{Path, ResourceType, Value};
+use crate::{
+    parse::{Path, ResourceType, Value},
+};
 
 /// An output.
 #[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
@@ -50,6 +55,8 @@ impl<'a> From<&'a str> for SinkKind {
 #[serde(rename_all = "kebab-case")]
 pub struct Model {
     pub model_file: ModelFile,
+    pub format: ModelFormat,
+    pub args: IndexMap<String, String>,
 }
 
 /// Where to load a model from.
@@ -288,3 +295,61 @@ impl Deref for ModelData {
 
     fn deref(&self) -> &Self::Target { &self.0 }
 }
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, serde::Serialize, serde::Deserialize,
+)]
+#[non_exhaustive]
+pub enum ModelFormat {
+    TensorFlowLite,
+    TensorFlow,
+    ONNX,
+}
+
+impl ModelFormat {
+    pub fn mimetype(self) -> &'static str {
+        match self {
+            ModelFormat::TensorFlowLite => hotg_rune_core::TFLITE_MIMETYPE,
+            ModelFormat::TensorFlow => hotg_rune_core::TF_MIMETYPE,
+            ModelFormat::ONNX => hotg_rune_core::ONNX_MIMETYPE,
+        }
+    }
+}
+
+impl Default for ModelFormat {
+    fn default() -> Self { ModelFormat::TensorFlowLite }
+}
+
+impl FromStr for ModelFormat {
+    type Err = UnknownFormatError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "tensorflow-lite" => Ok(ModelFormat::TensorFlowLite),
+            "tensorflow" => Ok(ModelFormat::TensorFlow),
+            "onnx" => Ok(ModelFormat::ONNX),
+            other => Err(UnknownFormatError {
+                format: other.to_string(),
+                expected: &["tensorflow-lite", "tensorflow", "onnx"],
+            }),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct UnknownFormatError {
+    pub format: String,
+    pub expected: &'static [&'static str],
+}
+
+impl Display for UnknownFormatError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "Expected the format to be one of {:?}, but found \"{}\"",
+            self.expected, self.format
+        )
+    }
+}
+
+impl Error for UnknownFormatError {}