+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - You can now specify a model's format by setting the `format` field under
   `args` to one of `tensorflow-lite`, `tensorflow`, or `onnx`
+  ([#367](https://github.com/hotg-ai/rune/pull/367))
   - Note that this will just ask the runtime to load a particular model, there
     is no guarantee it will be supported
 
@@ -20,6 +21,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The `hotg-rune-wasmer-runtime` and `hotg-rune-wasm3-runtime` crates have been
   merged into `hotg-rune-runtime`
+
+### Fixed
+
+- Warnings are no longer printed multiple times during `rune build`
+  ([#367](https://github.com/hotg-ai/rune/pull/367))
 
 ## [0.10.0] - 2021-10-25